+++ src/main/java/GaVisionUp/server/web/controller/exp/ExpBarController.java
@@ -20,13 +20,13 @@ public class ExpBarController {
 
     // 특정 사원의 경험치 바 조회
     @GetMapping("/{userId}")
-    public ResponseEntity<ExpBar> getExpBar(@PathVariable int userId) {
+    public ResponseEntity<ExpBar> getExpBar(@PathVariable Long userId) {
         return ResponseEntity.ok(expBarService.getExpBarByUserId(userId));
     }
 
     // 총 경험치에 새로 경험치 추가
     @PostMapping("/{userId}/add-exp")
-    public ResponseEntity<ExpBar> addExperience(@PathVariable int userId, @RequestParam int exp) {
+    public ResponseEntity<ExpBar> addExperience(@PathVariable Long userId, @RequestParam int exp) {
         return ResponseEntity.ok(expBarService.addExperience(userId, exp));
     }
 }
